Bugfix: (start!) opens web browser (again)
I introduced this bug when I refactored the start! function to take an
optional option map. I wanted to be able to restart everything without
launching new tabs for web browser.

But I didn't finish the refactoring. I introduced a `:browse? true`
argument for the start! function, but the default argument was `nil`,
not `true`.

Now the default argument is `true`.