Proposal: store, cohort and doc
Untill now, we have struggled to figure out whether functionality should
live in `mikrobloggeriet.cohort` or `mikrobloggeriet.doc`. A symptom of
this is cyclic dependencies. Does doc depend on cohort? Does cohort
depend on doc?

We propose to break this dependency cycle by keeping the abstractions
for cohort and doc narrow, and introducing a _store_. The store knows
about cohorts and documents.