`mblog create --no-edit`: Require less configuration (#43)
* improve `mblog create --no-edit`

Don't require a valid configured `:editor` in this case. The --no-edit
flag disables launching an editor, so it doesn't really require an
editor to be set.

* Fix: actually treat :edit right this time

It didn't work before, because `false` is falsey, and thus can't be used
in `or`. Should actually be correct now.

* Actually try making it work, take 2

When we've set `--no-edit`, we shouldn't run git commands!

(trying to run git commands without first having shelled out to an
editor would cause us to publish empty documents.)

* `mblog create`: code comment nitpicks

* `mblog create`: further code comment nitpicks

---------

Co-authored-by: Teodor Heggelund <git@teod.eu>