Refactor: be precise about cohort usage
Old cohort usage was only a specific type of cohort. Now that we know
there are _different kinds_ of cohorts, we can think precisely about
just that.

The goal is for `mikrobloggeriet.serve` to depend on
`mikrobloggeriet.cohort`, and `mikrobloggeriet.cohort` to depend on
`mikrobloggeriet.cohort.markdown` and `mikrobloggeriet.cohort.urlog`.