Work-in-progress analytics stunt
So ...

The fact that you're reading this commit message means you're quite dedicated. I
expect this commit only to live on a branch that will never see the light of
day.

This is an example of exactly the kind of programming workflow that I normally
discourage:

- Do a huge chunk of work
- Include breaking changes
- Don't commit working stuff as you go
- Finally, break the build
- And include random changes random places in the code.

From reading the diff and trying to recall (vaguely) what I intended about a
month ago (when I wrote the stuff, and didn't commit anything), I think my goal
was to push the information about which pages to do analytics for into the route
table, so that we can more easily present route information.

Right now, I don't feel inclined to continue that piece of work. Thus, off to a
branch it is.